+++ lib/cmds/ex_actions.py
@@ -196,7 +196,7 @@ class ExHelp(ViWindowCommandBase):
             if not tags_resources:
                 return nvim.message('tags file not found')
 
-            tags_matcher = re.compile('^([^\s]+)\s+([^\s]+)\s+(.+)$')
+            tags_matcher = re.compile('^([^\\s]+)\\s+([^\\s]+)\\s+(.+)$')
             tags_resource = load_resource(tags_resources[0])
             for line in tags_resource.split('\n'):
                 if line:
@@ -216,7 +216,7 @@ class ExHelp(ViWindowCommandBase):
             tag[0]) if r.startswith('Packages/NeoVintageous/res/doc/')]
 
         if not doc_resources:
-            return nvim.message('no help file found for %s' % tag[0])
+            return nvim.message('Sorry, help file "%s" not found' % tag[0])
 
         # TODO REFACTOR into reusable api
         def window_find_open_view(window, name):