+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed: `vintageousrc` mapping should not accept unescaped pipe characters
 * Fixed: Help syntax fixes
 * Fixed: Unimpaired toggles (documentation)